Started Implementing Floyd Warshall Algorithm
The code is able to find the right truck to get the materials to from the warehouse to the production line and gives the truck that path. The issue is that the Disjkstra algorithm is slow and having to use it so many times in a for loop makes it slow. Floyd Marshall would be faster because it just takes a while at the beginning and then you can get the paths quickly. The implementation needs some work because it is not properly returning the shortest path - we can ask takac tomorrow.